+++ src/main/java/io/leangen/graphql/annotations/types/GraphQLType.java
@@ -15,6 +15,4 @@ public @interface GraphQLType {
     String name() default "";
 
     String description() default "";
-
-    String[] fieldOrder() default {};
 }