+++ src/main/java/io/leangen/graphql/generator/mapping/common/ObjectTypeMapper.java
@@ -21,9 +21,7 @@ import io.leangen.graphql.util.GraphQLUtils;
 import java.lang.reflect.AnnotatedType;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.TreeMap;
 import java.util.stream.Collectors;
 
 import static graphql.schema.GraphQLInputObjectField.newInputObjectField;
@@ -38,7 +36,7 @@ public class ObjectTypeMapper extends CachingMapper<GraphQLObjectType, GraphQLIn
                 .name(typeName)
                 .description(buildContext.typeInfoGenerator.generateTypeDescription(javaType, buildContext.messageBundle));
 
-        List<GraphQLFieldDefinition> fields = getFields(javaType, buildContext, operationMapper);
+        List<GraphQLFieldDefinition> fields = getFields(typeName, javaType, buildContext, operationMapper);
         fields.forEach(typeBuilder::field);
 
         List<GraphQLOutputType> interfaces = getInterfaces(javaType, fields, buildContext, operationMapper);
@@ -87,11 +85,10 @@ public class ObjectTypeMapper extends CachingMapper<GraphQLObjectType, GraphQLIn
     }
 
     @SuppressWarnings("WeakerAccess")
-    protected List<GraphQLFieldDefinition> getFields(AnnotatedType javaType, BuildContext buildContext, OperationMapper operationMapper) {
-        List<GraphQLFieldDefinition> fields = buildContext.operationRegistry.getChildQueries(javaType).stream()
-                .map(childQuery -> operationMapper.toGraphQLField(childQuery, buildContext))
+    protected List<GraphQLFieldDefinition> getFields(String typeName, AnnotatedType javaType, BuildContext buildContext, OperationMapper operationMapper) {
+        return buildContext.operationRegistry.getChildQueries(javaType).stream()
+                .map(childQuery -> operationMapper.toGraphQLField(typeName, childQuery, buildContext))
                 .collect(Collectors.toList());
-        return sortFields(fields, buildContext.typeInfoGenerator.getFieldOrder(javaType, buildContext.messageBundle));
     }
 
     @SuppressWarnings("WeakerAccess")
@@ -106,21 +103,6 @@ public class ObjectTypeMapper extends CachingMapper<GraphQLObjectType, GraphQLIn
                 inter -> interfaces.add(operationMapper.toGraphQLType(inter, buildContext)));
 
         return interfaces;
-    }
-
-    private static List<GraphQLFieldDefinition> sortFields(List<GraphQLFieldDefinition> fields, String[] specifiedFieldOrder) {
-        Map<String, GraphQLFieldDefinition> fieldMap = new TreeMap<>();
-        for (GraphQLFieldDefinition field : fields) {
-            fieldMap.put(field.getName(), field);
-        }
-        List<GraphQLFieldDefinition> result = new ArrayList<>();
-        for (String name : specifiedFieldOrder) {
-            if (fieldMap.containsKey(name)) {
-                result.add(fieldMap.remove(name));
-            }
-        }
-        result.addAll(fieldMap.values());
-        return result;
     }
 
     @SuppressWarnings("WeakerAccess")