+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for `OffsetTime` [#260](https://github.com/leangen/graphql-spqr/issues/260)
 
 ### Changed
+- [Breaking] Upgraded to graphql-java v12 [#252](https://github.com/leangen/graphql-spqr/issues/252)
+- [Breaking] Field sorting no longer possible (due to graphql-java v12 upgrade)
 - Jackson and ClassGraph versions upgraded [#226](https://github.com/leangen/graphql-spqr/issues/226) [#227](https://github.com/leangen/graphql-spqr/issues/227)
 - Small improvements [a663162](https://github.com/leangen/graphql-spqr/commit/a663162430add139fbb706f2973cbea8e9d0df78)
 - `JavaDeprecationMappingConfig` accessible from `GeneratorConfiguration` [4aa8c5f](https://github.com/leangen/graphql-spqr/commit/4aa8c5fc539ef3b2837349df00ea318271e8f6cb) [7a34178](https://github.com/leangen/graphql-spqr/commit/7a341788cc2501232c77336bdb969f4c3fc551ef)